Restructure Component File System
* Move chat relevant components to the "Chat" folder

* course -> /Courses

* account -> /Account
