+++ App.js
@@ -5,18 +5,18 @@ import { createStackNavigator } from "@react-navigation/stack";
 import * as firebase from "firebase";
 import { Icon } from 'react-native-elements';
 import Home from "./components/Home";
-import Login from "./components/Login/Login";
+import Login from "./components/Account/Login/Login";
 import Register from "./components/Register/Register";
-import ChatPage from "./components/ChatPage";
-import ChatSelection from "./components/ChatSelection";
-import Settings from "./components/Settings";
-import UpdateProfilePicture from "./components/UpdateProfilePicture";
-import UpdatePassword from "./components/UpdatePassword";
-import Add from "./components/Add";
-import DM from "./components/DM";
+import ChatPage from "./components/Chat/ChatPage";
+import ChatSelection from "./components/Chat/ChatSelection";
+import ChatHeader from "./components/Chat/ChatHeader";
+import Settings from "./components/Account/Settings";
+import UpdateProfilePicture from "./components/Account/UpdateProfilePicture";
+import UpdatePassword from "./components/Account/UpdatePassword";
+import Add from "./components/Courses/Add";
+import DM from "./components/Chat/DM";
 import { StyleSheet, Button, Text, View } from "react-native";
-import Pins from "./components/Pins";
-import ChatHeader from "./components/ChatHeader";
+import Pins from "./components/Chat/Pins";
 import { TouchableOpacity } from "react-native-gesture-handler";
 
 const Stack = createStackNavigator();
