+++ debug.log
@@ -1,0 +1,3 @@
+[1023/095654.778:ERROR:filesystem_win.cc(129)] GetFileAttributes C:\Users\ov78\AppData\Local\Google\Chrome\User Data\Crashpad\attachments\39162ab0-1c24-418f-a87f-d07f82305691: The system cannot find the path specified. (0x3)
+[1023/095654.783:ERROR:filesystem_win.cc(129)] GetFileAttributes C:\Users\ov78\AppData\Local\Google\Chrome\User Data\Crashpad\attachments\7be3f84c-1595-4f98-ae01-411d36a5939d: The system cannot find the path specified. (0x3)
+[1023/095654.785:ERROR:directory_reader_win.cc(43)] FindFirstFile: The system cannot find the path specified. (0x3)
